GET list reject participants & Fix bugs of GET Event by id

+++ backend/Controllers/EventController.cs
@@ -61,10 +61,11 @@ namespace backend.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> GetNumberOfParticipants(string id, [FromQuery] string? status)
         {
+            if (!ObjectId.TryParse(id, out _)) return NotFound("Invalid ID");
             var e = await _eventRepository.Get(id);
             if (e == null)
             {
-                return BadRequest("Cannot find any Event from this _id");
+                return NotFound("Cannot find any Event from this _id");
             }
             if (status is "upcoming")
             {
@@ -84,6 +85,12 @@ namespace backend.Controllers
         [HttpGet("listParticipants/{id}")]
         public async Task<IActionResult> GetListParticipants(string id)
         {
+            if (!ObjectId.TryParse(id, out _)) return NotFound("Invalid ID");
+            var e = await _eventRepository.Get(id);
+            if (e == null)
+            {
+                return NotFound("Cannot find any Event from this _id");
+            }
             var result = new List<Donor>();
             var listDonor = await _donorRepository.Get();
             foreach (var donor in listDonor)
@@ -97,11 +104,6 @@ namespace backend.Controllers
                 result.Add(tempDonor);
             }
 
-            if (!result.Any())
-            {
-                return NotFound();
-            }
-
             var sortResult = result.OrderByDescending(d => long.Parse(d.transaction.dateDonated));
 
             return new JsonResult(sortResult);
@@ -110,16 +112,25 @@ namespace backend.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(string id, Event e)
         {
+            if (!ObjectId.TryParse(id, out _)) return NotFound("Invalid ID");
             var result = await _eventRepository.Update(id, e);
-            return new JsonResult(result);
+            if (!result)
+            {
+                return BadRequest();
+            }
+            return new JsonResult("Update Information Of Event Successfully");
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(string id)
         {
+            if (!ObjectId.TryParse(id, out _)) return NotFound("Invalid ID");
             var result = await _eventRepository.Delete(id);
-
-            return new JsonResult(result);
+            if (!result)
+            {
+                return BadRequest();
+            }
+            return new JsonResult("Delete Information Of Event Successfully");
         }
     }
 }
